all_modules.py
import numpy as np 
import pandas as pd
import re
from pathlib import Path
import time
from tqdm import tqdm
from string import punctuation
from nltk.corpus import stopwords
# sentiment 
from textblob import TextBlob
import tweepy
import matplotlib.pyplot as plt


import seaborn as sns


# And libraries for data transformation
import datetime

#words counter
from collections import Counter
# Libraries for wordcloud making and image importing
from wordcloud import WordCloud, ImageColorGenerator
from PIL import Image

# And libraries for data transformation
import datetime

#words counter
from collections import Counter
import re
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
import spacy
from nltk.corpus import stopwords
import re
from sklearn import preprocessing
from PIL import Image
# And libraries for data transformation
import datetime
#words counter
from collections import Counter
import re

from plotly.offline import iplot
import pandas as pd
import numpy as np
import itertools
import matplotlib.pyplot as plt
from sklearn.metrics import confusion_matrix
import pickle
import sklearn
from sklearn import linear_model
from afinn import Afinn
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer 

import json
import jsonlines
import numpy as np 
import pandas as pd
import os

from pathlib import Path

from settings import token
from settings import us_states
from settings import states
from settings import states_abb_dic
from settings import states_full_dic
from tqdm import tqdm

import time