requirements.txt
afinn==0.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
idna==2.9
importlib-metadata==1.6.0
joblib==0.15.1
jsonlines==1.2.0
kiwisolver==1.2.0
matplotlib==3.2.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.1.2
plac==1.1.3
plotly==4.7.1
preshed==3.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
spacy==2.2.4
srsly==1.0.2
textblob==0.15.3
thinc==7.4.0
threadpoolctl==2.0.0
tqdm==4.46.0
tweepy==3.8.0
urllib3==1.25.9
vaderSentiment==3.3.2
wasabi==0.6.0
wordcloud==1.7.0
zipp==3.1.0
